+++ src/main/kotlin/org/example/loja_das_coisas_api/store/model/StoreWithdraw.kt
@@ -3,16 +3,17 @@ package org.example.loja_das_coisas_api.store.model
 import jakarta.persistence.*
 import org.example.loja_das_coisas_api.shared.model.BaseEntity
 import org.example.loja_das_coisas_api.store.dto.StoreWithdrawDtoRes
+import java.math.BigDecimal
 
 @Entity
 @Table(name = "store_withdraws")
 class StoreWithdraw(
     @ManyToOne(fetch = FetchType.LAZY)
     var store: StoreProfile,
-    var amount: Int,
-    var fee: Int,
-    var feeAmount: Int,
-    var total: Int,
+    var amount: BigDecimal,
+    var fee: Double,
+    var feeAmount: BigDecimal,
+    var total: BigDecimal,
     @Enumerated(EnumType.STRING)
     var status: WithdrawStatus = WithdrawStatus.Pending,
 ): BaseEntity()
